+++ .dockerignore
@@ -10,3 +10,4 @@ hub/
 .git/
 .idea/
 **/*.iml
+hf_cache
