+++ examples/automatic-speech-recognition/whisper.py
@@ -16,15 +16,18 @@ from argparse import ArgumentParser
 from logging import getLogger
 from pathlib import Path
 
+import torch
 from huggingface_hub import login
 
 from optimum.nvidia import setup_logging
 from transformers import AutoProcessor
-import torch
+
 
 # Setup logging needs to happen before importing TRT ...
 setup_logging(False)
 LOGGER = getLogger(__name__)
+
+from datasets import load_dataset
 
 from optimum.nvidia.models.whisper import WhisperForConditionalGeneration
 from optimum.nvidia.utils.cli import (
@@ -32,7 +35,7 @@ from optimum.nvidia.utils.cli import (
     register_optimization_profiles_args,
     register_quantization_args,
 )
-from datasets import load_dataset
+
 
 if __name__ == "__main__":
     parser = ArgumentParser("🤗 TensorRT-LLM Whisper implementation")
@@ -55,18 +58,28 @@ if __name__ == "__main__":
         login(args.hub_token)
 
     torch_dtype = torch.float32
-    model = WhisperForConditionalGeneration.from_pretrained(args.model, torch_dtype=torch_dtype)
+    model = WhisperForConditionalGeneration.from_pretrained(
+        args.model, torch_dtype=torch_dtype
+    )
     model.save_pretrained(args.output)
     print(f"TRTLLM engines have been saved at {args.output}.")
 
     model = WhisperForConditionalGeneration.from_pretrained(args.output)
-    processor = AutoProcessor.from_pretrained(args.model)  # TODO: save processor in args.output (like ORTModel)
+    processor = AutoProcessor.from_pretrained(
+        args.model
+    )  # TODO: save processor in args.output (like ORTModel)
 
-    data = load_dataset("hf-internal-testing/librispeech_asr_dummy", "clean", split="validation")
+    data = load_dataset(
+        "hf-internal-testing/librispeech_asr_dummy", "clean", split="validation"
+    )
 
-    inputs = processor(data[0]["audio"]["array"], return_tensors="pt", sampling_rate=data[0]["audio"]["sampling_rate"]).to("cuda")
+    inputs = processor(
+        data[0]["audio"]["array"],
+        return_tensors="pt",
+        sampling_rate=data[0]["audio"]["sampling_rate"],
+    ).to("cuda")
     input_features = inputs.input_features
-    
+
     input_features = input_features.to(torch_dtype)
     generated_ids = model.generate(inputs=input_features)
