+++ examples/quantization.py
@@ -17,10 +17,9 @@ from argparse import ArgumentParser
 from logging import getLogger
 from pathlib import Path
 
-from transformers import AutoTokenizer
-
 from optimum.nvidia import AutoModelForCausalLM, setup_logging
 from optimum.nvidia.quantization import AutoQuantizationConfig
+from transformers import AutoTokenizer
 
 
 # Setup logging needs to happen before importing TRT ...
